Remove Tile and TileHeap from grid.py
