+++ grid.py
@@ -1,157 +1,8 @@
 import math
+from Tile import *
 
-
-class Tile:
-    def __init__(self, x, y, row, col, isObstacle=False, isBloated=False, isKnown=False):
-        """
-        Initialize a tile centered at x coordinate [x] and y coordinate [y].
-        If [isObstacle] is True, the tile is initialized as an obstacle, else Tile is
-        marked as free space, [isObstacle] is False by default.
-        """
-        self.x = x
-        self.y = y
-        self.row = row
-        self.col = col
-        self.isObstacle = isObstacle
-        self.isBloated = isBloated
-
-
-class TileHeap:
-    def __init__(self):
-        """
-        Initializes an empty heap of tiles.
-        """
-        self.data = []
-        self.cost_map = {}  # to keep track of each tiles cost + heuristic
-        self.idx_map = {}  # to keep track of each tiles index in data
-
-    def comparator(self, pos1, pos2):
-        """
-        given index [pos1] and index [pos2] of the heap, returns 0 if tile at
-        [pos1] has same cost + heuristic as the tile at [pos2]; returns -1 if
-        tile at [pos1] has greater cost + heuristic than tile at [pos2] and returns
-        1 if tile at [pos1] has smaller cost + heuristic than tile at [pos2].
-        """
-        tile1 = self.data[pos1]
-        tile2 = self.data[pos2]
-        data1 = self.cost_map[tile1][0] + self.cost_map[tile1][1]
-        data2 = self.cost_map[tile2][0] + self.cost_map[tile2][1]
-        if data1 == data2:
-            return 0
-        elif data1 > data2:
-            return -1
-        else:
-            return 1
-
-    def push(self, elt, cost, heuristic):
-        """
-        Push tile [elt] on to the heap with cost [cost] and heuristic estimate
-        [heuristic]
-
-        assumes: [elt] is a Tile object, [cost] and [heuristic] are floats
-        """
-        if elt in self.idx_map:
-            return None
-
-        pos = len(self.data)
-        self.idx_map[elt] = pos
-        self.cost_map[elt] = [cost, heuristic]
-        self.data.append(elt)
-        self._bubble_up(pos)
-
-    def pop(self):
-        """
-        Pop tile with minimum cost + heuristic from heap and return a tuple of tile
-        and cost to get to the tile (in that order), returns None if heap is empty.
-        """
-        if self.isEmpty():
-            return None
-
-        elt = self.data[0]
-        cost = self.cost_map[elt][0]
-        del self.idx_map[elt]
-        del self.cost_map[elt]
-        if len(self.data) == 1:
-            self.data = []
-        else:
-            last = self.data.pop()
-            self.data[0] = last
-            self.idx_map[last] = 0
-            self._bubble_down(0)
-
-        return (elt, cost)
-
-    def _swap(self, pos1, pos2):
-        """
-        swap tile at [pos1] in heap with tile at [pos2] in heap.
-        """
-        elt1 = self.data[pos1]
-        elt2 = self.data[pos2]
-        self.idx_map[elt1], self.idx_map[elt2] = pos2, pos1
-        self.data[pos1], self.data[pos2] = elt2, elt1
-
-    def _bubble_up(self, pos):
-        """
-        helper function for [push] and [updatePriority]
-        """
-        parent = (pos - 1) // 2
-        while pos > 0 and self.comparator(pos, parent) > 0:
-            self._swap(pos, parent)
-            pos = parent
-            parent = (pos - 1) // 2
-
-    def _biggerChild(self, pos):
-        """
-        helper function for [pop], returns the child of element at [pos] with the
-        highest priority
-        """
-        c = 2 * pos + 2
-        if c >= len(self.data) or self.comparator(c - 1, c) > 0:
-            c = c - 1
-        return c
-
-    def _bubble_down(self, pos):
-        """
-        helper function for [pop] and [updatePriority]
-        """
-        child = self._biggerChild(pos)
-        while (child < len(self.data) and self.comparator(pos, child) < 0):
-            self._swap(pos, child)
-            pos = child
-            child = self._biggerChild(pos)
-
-    def updatePriority(self, elt, new_cost):
-        """
-        updates tile [elt] in the heap with cost [new_cost].
-        Assumes [elt] is already in the heap
-
-        Note: [new_cost] should not include heuristic estimate
-        """
-        self.cost_map[elt][0] = new_cost
-        pos = self.idx_map[elt]
-        self._bubble_up(pos)
-        self._bubble_down(pos)
-
-    def isEmpty(self):
-        """
-        Returns True if heap is empty, else returns False
-        """
-        return False if self.data else True
-
-    def mem(self, elt):
-        """
-        returns True if [elt] is in the heap, else returns False
-        """
-        return elt in self.idx_map
-
-    def getCost(self, elt):
-        """
-        Returns the cost from start to tile [elt] if [elt] is in heap, else returns
-        None
-        """
-        if not self.mem(elt):
-            return
-        return self.cost_map[elt][0]
+ir_mappings_top = {}
+ir_mappings_bot = {}
 
 
 class Grid:
@@ -172,74 +23,74 @@ class Grid:
         self.num_cols = num_cols
         # TODO change center pos
 
-    def updateGrid(self, x, y, sensorDataTop, sensorDataBot, lidarData, path):
-        """
-        Marks tiles of grid as occuppied using IR sensor data [sensorData] measured
-        from x position [x] and y position [y].
-        """
-        # returner is a variable to keep track of whether
-        # A-star needs to be re-run
-        returner = False
-        for i in range(len(sensorDataTop)):
-            angle = ir_mappings_top[i]
-            distance = sensorDataTop[i]
-            if distance != -1:
-                x_obst = x + radius + distance * math.cos(angle)
-                # y_obst = y + radius + distance * math.sin(angle)
-                y_obst = y + radius - distance * \
-                         math.sin(angle)  # upper left origin
-                col = self._get_idx(x_obst, False)
-                row = self._get_idx(y_obst, True)
-                if (self.grid[row][col].isObstacle == False):
-                    if row > len(self.grid) or col > len(self.grid[0]):
-                        # TODO handle offgrid case
-                        return
-                    if (self.grid[row][col] in path):
-                        returner = True
-                    self.grid[row][col].isObstacle = True
-                    if (self.bloat_tile(row, col, path) == True):
-                        returner = True
-
-        for i in range(len(sensorDataBot)):
-            angle = ir_mappings_bot[i]
-            distance = sensorDataBot[i]
-            if distance != -1:
-                x_obst = x + radius + distance * math.cos(angle)
-                # y_obst = y + radius + distance * math.sin(angle)
-                y_obst = y + radius - distance * \
-                         math.sin(angle)  # upper left origin
-                col = self._get_idx(x_obst, False)
-                row = self._get_idx(y_obst, True)
-                if (self.grid[row][col].isObstacle == False):
-                    if row > len(self.grid) or col > len(self.grid[0]):
-                        # TODO handle offgrid case
-                        return
-                    if (self.grid[row][col] in path):
-                        returner = True
-                    self.grid[row][col].isObstacle = True
-                    if (self.bloat_tile(row, col, radius) == True):
-                        returner = True
-
-        for i in lidarData:
-            angle = i[0]
-            distance = i[1]
-            if distance != -1:
-                x_obst = x + radius + distance * math.cos(angle)
-                # y_obst = y + radius + distance * math.sin(angle)
-                y_obst = y + radius - distance * \
-                         math.sin(angle)  # upper left origin
-                col = self._get_idx(x_obst, False)
-                row = self._get_idx(y_obst, True)
-                if (self.grid[row][col].isObstacle == False):
-                    if row > len(self.grid) or col > len(self.grid[0]):
-                        # TODO handle offgrid case
-                        return
-                    if (self.grid[row][col] in path):
-                        returner = True
-                    self.grid[row][col].isObstacle = True
-                    if (self.bloat_tile(row, col, radius) == True):
-                        returner = True
-            return returner
+    # def updateGrid(self, x, y, sensorDataTop, sensorDataBot, lidarData, path):
+    #     """
+    #     Marks tiles of grid as occuppied using IR sensor data [sensorData] measured
+    #     from x position [x] and y position [y].
+    #     """
+    #     # returner is a variable to keep track of whether
+    #     # A-star needs to be re-run
+    #     returner = False
+    #     for i in range(len(sensorDataTop)):
+    #         angle = ir_mappings_top[i]
+    #         distance = sensorDataTop[i]
+    #         if distance != -1:
+    #             x_obst = x + robot_radius + distance * math.cos(angle)
+    #             # y_obst = y + radius + distance * math.sin(angle)
+    #             y_obst = y + robot_radius - distance * \
+    #                      math.sin(angle)  # upper left origin
+    #             col = self._get_idx(x_obst, False)
+    #             row = self._get_idx(y_obst, True)
+    #             if (self.grid[row][col].isObstacle == False):
+    #                 if row > len(self.grid) or col > len(self.grid[0]):
+    #                     # TODO handle offgrid case
+    #                     return
+    #                 if (self.grid[row][col] in path):
+    #                     returner = True
+    #                 self.grid[row][col].isObstacle = True
+    #                 if (self.bloat_tile(row, col, path) == True):
+    #                     returner = True
+    #
+    #     for i in range(len(sensorDataBot)):
+    #         angle = ir_mappings_bot[i]
+    #         distance = sensorDataBot[i]
+    #         if distance != -1:
+    #             x_obst = x + robot_radius + distance * math.cos(angle)
+    #             # y_obst = y + radius + distance * math.sin(angle)
+    #             y_obst = y + robot_radius - distance * \
+    #                      math.sin(angle)  # upper left origin
+    #             col = self._get_idx(x_obst, False)
+    #             row = self._get_idx(y_obst, True)
+    #             if (self.grid[row][col].isObstacle == False):
+    #                 if row > len(self.grid) or col > len(self.grid[0]):
+    #                     # TODO handle offgrid case
+    #                     return
+    #                 if (self.grid[row][col] in path):
+    #                     returner = True
+    #                 self.grid[row][col].isObstacle = True
+    #                 if (self.bloat_tile(row, col, robot_radius) == True):
+    #                     returner = True
+    #
+    #     for i in lidarData:
+    #         angle = i[0]
+    #         distance = i[1]
+    #         if distance != -1:
+    #             x_obst = x + robot_radius + distance * math.cos(angle)
+    #             # y_obst = y + radius + distance * math.sin(angle)
+    #             y_obst = y + robot_radius - distance * \
+    #                      math.sin(angle)  # upper left origin
+    #             col = self._get_idx(x_obst, False)
+    #             row = self._get_idx(y_obst, True)
+    #             if (self.grid[row][col].isObstacle == False):
+    #                 if row > len(self.grid) or col > len(self.grid[0]):
+    #                     # TODO handle offgrid case
+    #                     return
+    #                 if (self.grid[row][col] in path):
+    #                     returner = True
+    #                 self.grid[row][col].isObstacle = True
+    #                 if (self.bloat_tile(row, col, robot_radius) == True):
+    #                     returner = True
+    #         return returner
 
     def updateGridLidar(self, x, y, lidarData, radius, bloat_factor, pathSet, fullGrid):
         """updates the grid based on the lidarData passed in
